fix(ci): make PostgreSQL Concurrency workflow actually pass, fixing real bugs it uncovered
The workflow has never been green. Root causes, in order of discovery:

1. npm ci could not run: package-lock.json was gitignored and the
   @release-it/conventional-changelog@^8 peer range (release-it ^17)
   conflicted with release-it ^19. Upgrade the plugin to ^10 (peer
   ^18||^19, changelog generation verified) and commit the lockfile.

2. 'Model manifest mismatch' in all multi-process tests: the migration
   manifest hashes computation callbacks via Function.prototype.toString().
   The vitest parent (vite/esbuild) and the tsx-launched workers produced
   different function text for the same source. Workers now run through
   vite-node with the vitest config (same transform pipeline), and the
   model fixtures are extracted into a single shared module
   (postgresConcurrencyFixtures.ts) imported by both spec and worker,
   since vite SSR import rewriting also makes per-file copies hash
   differently.

3. One test used the outdated Custom API: incrementalCompute now requires
   planIncremental or incrementalDataDeps; declare incrementalDataDeps: [].

Real framework bugs surfaced once the tests finally ran:

4. PostgreSQLDB.setupRecordSequences rolled sequences BACK to the
   committed MAX(id) whenever a process attached via setup(false).
   Concurrent writers had already consumed higher sequence values, so the
   rollback caused duplicate ids and corrupted incremental aggregates
   (reproduced: 4 workers x 12 creates lost sum increments). setval is now
   guarded to only advance, never rewind.

5. PostgreSQLDB.open() unconditionally created a new Pool, orphaning the
   pool created by openForSchemaRead() on the setup(false) path. Orphaned
   idle connections later killed by DROP DATABASE ... WITH (FORCE) raised
   process-level uncaught 'error' events (vitest exit 1 despite passing
   tests). open() now reuses the existing pool, and pools get the
   standard 'error' listener node-postgres requires for idle clients.

6. 'number' properties mapped to INT on PostgreSQL/PGLite/MySQL, so any
   legitimately fractional value (e.g. built-in Average = sum/count)
   failed with 'invalid input syntax for type integer'. JS numbers are
   double-precision floats; map them to DOUBLE PRECISION / DOUBLE.

7. The concurrency spec shared its database name with
   postgresqlDataConstraints.spec.ts; parallel vitest files DROP ... WITH
   (FORCE) each other mid-test. The spec now uses a dedicated
   ${INTERAQT_POSTGRES_DATABASE}_concurrency database like the other
   postgres specs.

Verified locally against PostgreSQL 16: npm ci, npm run check,
test:postgres-concurrency (16/16, 3 consecutive runs, exit 0, no
unhandled errors), test:postgres-scoped-sequence (2/2), and the full
suite twice with postgres enabled (123 files / 1647 tests, all green).
