+++ .gitignore
@@ -28,7 +28,6 @@ lib
 typings
 coverage
 .next
-package-lock.json
 temp
 pgdata
 test-create.sqlite
