+++ package.json
@@ -4,7 +4,7 @@
     "@electric-sql/pglite": "^0.3.2",
     "@interaqt/uuidv7": "^1.0.1",
     "@microsoft/api-extractor": "^7.52.1",
-    "@release-it/conventional-changelog": "^8.0.2",
+    "@release-it/conventional-changelog": "^10.0.6",
     "@types/better-sqlite3": "^7.6.12",
     "@types/node": "^22.13.10",
     "@types/pg": "^8.11.11",
