+++ tests/runtime/helpers/postgresConcurrencyWorker.ts
@@ -1,31 +1,20 @@
 import {
-  Action,
-  Any,
-  Average,
   Controller,
-  Count,
-  Custom,
-  ComputationResult,
-  Dictionary,
-  Entity,
-  Every,
-  Interaction,
-  InteractionEventEntity,
   KlassByName,
   MatchExp,
   MonoSystem,
-  Payload,
-  PayloadItem,
-  Property,
-  Relation,
-  StateMachine,
-  StateNode,
-  StateTransfer,
-  Summation,
-  Transform,
-  WeightedSummation,
 } from 'interaqt';
 import { PostgreSQLDB } from '@drivers';
+// CAUTION 模型定义必须与 vitest 父进程共享同一个模块（并通过同一转换管线执行），
+//  否则 migration manifest 的函数文本哈希会不一致，setup(false) 会报 "Model manifest mismatch"。
+import {
+  createAsyncReturnFixture,
+  createFullReplaceFixture,
+  createGlobalSummationFixture,
+  createPropertyAggregateFixture,
+  createStateMachineFixture,
+  createTransformFixture,
+} from './postgresConcurrencyFixtures.js';
 
 const database = process.env.INTERAQT_POSTGRES_DATABASE!;
 const mode = process.env.INTERAQT_POSTGRES_WORKER_MODE || 'global-summation';
@@ -39,257 +28,6 @@ const dbOptions = {
   user: process.env.PGUSER,
   password: process.env.PGPASSWORD,
 };
-
-function createGlobalSummationFixture() {
-  const counterEntity = Entity.create({
-    name: 'PgAtomicCounter',
-    properties: [Property.create({ name: 'value', type: 'number' })],
-  });
-
-  const totalDictionary = Dictionary.create({
-    name: 'pgAtomicTotal',
-    type: 'number',
-    collection: false,
-    computation: Summation.create({
-      record: counterEntity,
-      attributeQuery: ['value'],
-    }),
-  });
-
-  return { entities: [counterEntity], relations: [], dict: [totalDictionary] };
-}
-
-function createPropertyAggregateFixture() {
-  const User = Entity.create({
-    name: 'PgAggregateUser',
-    properties: [Property.create({ name: 'name', type: 'string' })],
-  });
-  const Order = Entity.create({
-    name: 'PgAggregateOrder',
-    properties: [
-      Property.create({ name: 'amount', type: 'number' }),
-      Property.create({ name: 'weight', type: 'number' }),
-    ],
-  });
-
-  User.properties.push(
-    Property.create({
-      name: 'orderCount',
-      type: 'number',
-      computation: Count.create({ property: 'orders' }),
-    }),
-    Property.create({
-      name: 'orderTotal',
-      type: 'number',
-      computation: Summation.create({ property: 'orders', attributeQuery: ['amount'] }),
-    }),
-    Property.create({
-      name: 'orderAverage',
-      type: 'number',
-      computation: Average.create({ property: 'orders', attributeQuery: ['amount'] }),
-    }),
-    Property.create({
-      name: 'allPositive',
-      type: 'boolean',
-      computation: Every.create({
-        property: 'orders',
-        attributeQuery: ['amount'],
-        callback: (order: any) => order.amount > 0,
-      }),
-    }),
-    Property.create({
-      name: 'hasLargeOrder',
-      type: 'boolean',
-      computation: Any.create({
-        property: 'orders',
-        attributeQuery: ['amount'],
-        callback: (order: any) => order.amount >= 1000,
-      }),
-    }),
-    Property.create({
-      name: 'weightedTotal',
-      type: 'number',
-      computation: WeightedSummation.create({
-        property: 'orders',
-        attributeQuery: ['amount', 'weight'],
-        callback: (order: any) => ({ value: order.amount || 0, weight: order.weight || 0 }),
-      }),
-    })
-  );
-
-  const UserOrder = Relation.create({
-    name: 'PgAggregateUserOrder',
-    source: User,
-    sourceProperty: 'orders',
-    target: Order,
-    targetProperty: 'buyer',
-    type: '1:n',
-  });
-
-  return { entities: [User, Order], relations: [UserOrder], dict: [] };
-}
-
-function createStateMachineFixture() {
-  const Order = Entity.create({
-    name: 'PgStateOrder',
-    properties: [Property.create({ name: 'title', type: 'string' })],
-  });
-  const Approve = Interaction.create({
-    name: 'pgApproveOrder',
-    action: Action.create({ name: 'pgApproveOrder' }),
-    payload: Payload.create({
-      items: [PayloadItem.create({ name: 'order', type: 'Entity', base: Order, isRef: true })],
-    }),
-  });
-  const Reject = Interaction.create({
-    name: 'pgRejectOrder',
-    action: Action.create({ name: 'pgRejectOrder' }),
-    payload: Payload.create({
-      items: [PayloadItem.create({ name: 'order', type: 'Entity', base: Order, isRef: true })],
-    }),
-  });
-  const pending = StateNode.create({ name: 'pending' });
-  const approved = StateNode.create({ name: 'approved' });
-  const rejected = StateNode.create({ name: 'rejected' });
-  Order.properties.push(Property.create({
-    name: 'status',
-    type: 'string',
-    computation: StateMachine.create({
-      states: [pending, approved, rejected],
-      initialState: pending,
-      transfers: [
-        StateTransfer.create({
-          trigger: { recordName: InteractionEventEntity.name, type: 'create', record: { interactionName: Approve.name } },
-          current: pending,
-          next: approved,
-          computeTarget: (event: any) => ({ id: event.record.payload.order.id }),
-        }),
-        StateTransfer.create({
-          trigger: { recordName: InteractionEventEntity.name, type: 'create', record: { interactionName: Reject.name } },
-          current: pending,
-          next: rejected,
-          computeTarget: (event: any) => ({ id: event.record.payload.order.id }),
-        }),
-      ],
-    }),
-  }));
-
-  return { entities: [Order], relations: [], eventSources: [Approve, Reject], approve: Approve, reject: Reject };
-}
-
-function createTransformFixture() {
-  const Source = Entity.create({
-    name: 'PgTransformSource',
-    properties: [Property.create({ name: 'items', type: 'number' })],
-  });
-  const Derived = Entity.create({
-    name: 'PgTransformDerived',
-    properties: [Property.create({ name: 'idx', type: 'number' })],
-    computation: Transform.create({
-      record: Source,
-      attributeQuery: ['*'],
-      callback: (source: any) => ({ idx: source.items || 0 }),
-    }),
-  });
-
-  return { entities: [Source, Derived], relations: [], dict: [] };
-}
-
-function createAsyncReturnFixture() {
-  const Source = Entity.create({
-    name: 'PgAsyncWorkerSource',
-    properties: [Property.create({ name: 'value', type: 'number' })],
-  });
-  const AddSource = Interaction.create({
-    name: 'pgAddAsyncWorkerSource',
-    action: Action.create({ name: 'pgAddAsyncWorkerSource' }),
-    payload: Payload.create({
-      items: [PayloadItem.create({ name: 'source', type: 'Entity', base: Source })],
-    }),
-  });
-  AddSource.resolve = async function(this: Controller, event: any) {
-    return this.system.storage.create('PgAsyncWorkerSource', event.payload.source);
-  };
-  const total = Dictionary.create({
-    name: 'pgAsyncWorkerTotal',
-    type: 'number',
-    collection: false,
-    computation: Custom.create({
-      name: 'PgAsyncWorkerTotal',
-      dataDeps: {
-        sources: { type: 'records', source: Source, attributeQuery: ['value'] },
-      },
-      compute: async () => ComputationResult.async({ freshnessKey: 'pg-async-worker' }),
-      asyncReturn: async (result: number) => result,
-    }),
-  });
-
-  return { entities: [Source], relations: [], eventSources: [AddSource], dict: [total], addSource: AddSource };
-}
-
-function createFullReplaceFixture() {
-  const Trigger = Entity.create({
-    name: 'PgReplaceTrigger',
-    properties: [Property.create({ name: 'value', type: 'number' })],
-  });
-  const AddTrigger = Interaction.create({
-    name: 'pgAddReplaceTrigger',
-    action: Action.create({ name: 'pgAddReplaceTrigger' }),
-    payload: Payload.create({
-      items: [PayloadItem.create({ name: 'trigger', type: 'Entity', base: Trigger })],
-    }),
-  });
-  AddTrigger.resolve = async function(this: Controller, event: any) {
-    return this.system.storage.create('PgReplaceTrigger', event.payload.trigger);
-  };
-  const Result = Entity.create({
-    name: 'PgReplaceResult',
-    properties: [Property.create({ name: 'value', type: 'number' })],
-    computation: Custom.create({
-      name: 'PgReplaceResultComputation',
-      dataDeps: {
-        triggers: { type: 'records', source: Trigger, attributeQuery: ['value'] },
-      },
-      compute: async (dataDeps: any) => (dataDeps.triggers || []).map((trigger: any) => ({ value: trigger.value })),
-    }),
-  });
-  const Left = Entity.create({
-    name: 'PgReplaceLeft',
-    properties: [Property.create({ name: 'name', type: 'string' })],
-  });
-  const Right = Entity.create({
-    name: 'PgReplaceRight',
-    properties: [Property.create({ name: 'name', type: 'string' })],
-  });
-  const ResultRelation = Relation.create({
-    name: 'PgReplaceRelation',
-    source: Left,
-    sourceProperty: 'replaceLinks',
-    target: Right,
-    targetProperty: 'replaceLinkedBy',
-    type: 'n:n',
-    properties: [Property.create({ name: 'value', type: 'number' })],
-    computation: Custom.create({
-      name: 'PgReplaceRelationComputation',
-      dataDeps: {
-        triggers: { type: 'records', source: Trigger, attributeQuery: ['value'] },
-        lefts: { type: 'records', source: Left, attributeQuery: ['id'] },
-        rights: { type: 'records', source: Right, attributeQuery: ['id'] },
-      },
-      compute: async (dataDeps: any) => {
-        const lefts = dataDeps.lefts || [];
-        const rights = dataDeps.rights || [];
-        return (dataDeps.triggers || []).slice(0, Math.min(lefts.length, rights.length)).map((trigger: any, index: number) => ({
-          source: { id: lefts[index].id },
-          target: { id: rights[index].id },
-          value: trigger.value,
-        }));
-      },
-    }),
-  });
-
-  return { Trigger, AddTrigger, Result, Left, Right, ResultRelation, entities: [Trigger, Result, Left, Right], relations: [ResultRelation], eventSources: [AddTrigger] };
-}
 
 function delay(ms: number) {
   return new Promise(resolve => setTimeout(resolve, ms));
@@ -378,7 +116,7 @@ async function runStateMachineWorker() {
   const controller = new Controller({ system, ...fixture });
   await controller.setup(false);
 
-  await controller.dispatch(action === 'approve' ? fixture.approve : fixture.reject, {
+  await controller.dispatch(action === 'approve' ? fixture.Approve : fixture.Reject, {
     user: { id: `pg-state-worker-${workerIndex}` },
     payload: { order: { id: orderId } },
   });
